requirements.txt
# Core dependencies
redis[hiredis]>=5.0.0
aioredis>=2.0.1
asyncio>=3.4.3
pydantic>=2.5.0
pydantic-settings>=2.1.0
ulid-py>=1.1.0
PyYAML>=6.0

# AI/ML dependencies
openai>=1.12.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-redis>=0.44b0
opentelemetry-exporter-otlp>=1.22.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# CLI
click>=8.1.0
tabulate>=0.9.0
